Refactor CalendarItem file structure
