+++ src/nodes/events-calendar/CalendarItem.ts
@@ -1,34 +1,4 @@
-// Define type aliases
-type WithDateTime = {
-    dateTime: string;
-};
-
-type WithDate = {
-    date: string;
-};
-
-// Create a union type
-type DateOrDateTime = WithDateTime | WithDate;
-
-// Type guard for WithDateTime
-function isWithDateTime(obj: any): obj is WithDateTime {
-    return obj && typeof obj.dateTime === 'string';
-}
-
-// Type guard for WithDate
-function isWithDate(obj: any): obj is WithDate {
-    return obj && typeof obj.date === 'string';
-}
-
-// Function to handle the conversion at runtime
-function toDate(obj: DateOrDateTime): Date {
-    if (isWithDateTime(obj)) {
-        return new Date(obj.dateTime);
-    } else if (isWithDate(obj)) {
-        return new Date(obj.date);
-    }
-    throw new Error('Invalid object');
-}
+import { DateOrDateTime, toDate } from './helpers';
 
 export interface ICalendarItem {
     start: DateOrDateTime;
